requirements.txt
python-telegram-bot==20.0
google-api-python-client==2.86.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
